% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuge-package.R
\docType{package}
\name{refuge-package}
\alias{refuge}
\alias{refuge-package}
\title{refuge}
\description{
\href{Refuge Restrooms}{https://www.refugerestrooms.org/}
"is a web application that seeks to provide safe restroom access for
transgender, intersex, and gender nonconforming individuals." The web
application includes a public API, which this package allows access to.
The API does not require users to sign up or use a key.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evanodell/refuge}
  \item Report bugs at \url{https://github.com/evanodell/refuge/issues}
}

}
\author{
\strong{Maintainer}: Evan Odell \email{evanodell91@gmail.com}

}
\keyword{internal}
