% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_bathrooms.R, R/search.R
\name{rfg_all_restrooms}
\alias{rfg_all_restrooms}
\alias{rfg_all_restrooms}
\title{All restrooms}
\usage{
rfg_all_restrooms(search, accessible = FALSE, unisex = FALSE,
  verbose = TRUE)

rfg_all_restrooms(search, accessible = FALSE, unisex = FALSE,
  verbose = TRUE)
}
\arguments{
\item{search}{A string to search for. Note that the API does not accept
wildcard searchs, searches all fields in all available records, and only
returns exact (case-insensitive) matches.}

\item{accessible}{If \code{TRUE}, only returns accessible restrooms. Defaults to \code{FALSE}.}

\item{unisex}{If \code{TRUE}, only returns unisex restrooms. Defaults to \code{FALSE}.}

\item{verbose}{If \code{TRUE}, prints query progress. Defaults to \code{TRUE}.}
}
\description{
All restrooms, ordered by descending date.

Restroom record search
}
\examples{




}
