% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuge-package.R
\docType{package}
\name{refuge-package}
\alias{refuge}
\alias{refuge-package}
\title{refuge}
\description{


The API does not require an API key or user registration.
Requests are made over HTTPS, and returned in JSON format. All request
results are returned as tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.evanodell.com/refuge}
  \item Report bugs at \url{https://github.com/evanodell/refuge/issues}
}

}
\author{
\strong{Maintainer}: Evan Odell \email{evanodell91@gmail.com} (\href{https://orcid.org/0000-0003-1845-808X}{ORCID})

}
\keyword{internal}
