% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refuge-package.R
\docType{package}
\name{refuge-package}
\alias{refuge}
\alias{refuge-package}
\title{refuge}
\description{
\href{Refuge Restrooms}{https://www.refugerestrooms.org/}
"is a web application that seeks to provide safe restroom access for
transgender, intersex, and gender nonconforming individuals." The web
application includes a public API, which this package allows access to.
The API does not require users to sign up or use a key. For further details
on the API or the organisation behind it,
see the \href{API docs}{https://www.refugerestrooms.org/api/docs/}, or the
\href{Refuge Restrooms homepage}{https://www.refugerestrooms.org/}.

The API does not require an API key or user registration.
Requests are made over HTTPS, and returned in JSON format. All request
results are returned as tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evanodell/refuge}
  \item Report bugs at \url{https://github.com/evanodell/refuge/issues}
}

}
\author{
\strong{Maintainer}: Evan Odell \email{evanodell91@gmail.com} (0000-0003-1845-808X)

}
\keyword{internal}
